+++ lunatrace/cli/fx/grypefx/store/gqlstorefx/module.go
@@ -16,6 +16,7 @@ package gqlstorefx
 
 import (
 	"context"
+	"errors"
 	"time"
 
 	"github.com/Khan/genqlient/graphql"
@@ -24,7 +25,6 @@ import (
 	"go.uber.org/fx"
 
 	"github.com/lunasec-io/lunasec/lunatrace/cli/gql"
-	"github.com/lunasec-io/lunasec/lunatrace/cli/gql/types"
 )
 
 const SchemaVersion = 42069
@@ -48,33 +48,11 @@ func (g *gqlStore) GetID() (*v3.ID, error) {
 	}, nil
 }
 
-func mapNamespace(namespace string) types.PackageManager {
-	//TODO implement me
-	panic("implement me")
-}
-
 // GetVulnerability retrieves one or more vulnerabilities given a namespace and package name.
 func (g *gqlStore) GetVulnerability(namespace, name string) ([]v3.Vulnerability, error) {
 	vulns, err := gql.GetVulnerability(context.TODO(), g.d.GQLClient, name, mapNamespace(namespace))
 	//TODO implement me
 	panic("implement me")
-}
-
-// n2z converts nil pointers to the zero value of their type.
-func n2z[T any](test *T) T {
-	var result T
-	if test == nil {
-		return result
-	}
-	return *test
-}
-
-func mapURLs(urls []*gql.GetVulnerabilityMetadataVulnerability_by_pkVulnerabilityReferencesVulnerability_reference) []string {
-	out := make([]string, len(urls))
-	for i, ou := range urls {
-		out[i] = ou.Url
-	}
-	return out
 }
 
 // GetVulnerabilityMetadata retrieves metadata for the given vulnerability ID relative to a specific record source.
@@ -84,6 +62,12 @@ func (g *gqlStore) GetVulnerabilityMetadata(id, namespace string) (*v3.Vulnerabi
 		return nil, err
 	}
 	meta, err := gql.GetVulnerabilityMetadata(context.TODO(), g.d.GQLClient, uid)
+	if err != nil {
+		return nil, err
+	}
+	if meta.Vulnerability_by_pk == nil {
+		return nil, errors.New("vulnerability not found by ID")
+	}
 
 	return &v3.VulnerabilityMetadata{
 		ID:           meta.Vulnerability_by_pk.Id.String(),
